refactor: complete workspace migration from cli/ to crates/
- Delete old cli/ directory (monolithic structure)
- Move tests to crates/agent-tui/tests/
- Expand handlers.rs with migrated command logic
- Update CI workflows for workspace structure
- Update build scripts and package.json paths
