--- .ralph/progress.txt
@@ -1,65 +1,0 @@
-# Workspace Migration Progress Log
-# Each iteration appends learnings here
-
----
-[2025-01-22 21:24] Initialized workspace structure
-- Created root Cargo.toml with workspace manifest
-- Created crates/ directory with all 6 crates
-- Migrated source files to respective crates
-- Workspace builds successfully
-
-[2025-01-22 21:47] RALPH setup fixed
-- Added missing SPEC.md, status.json, progress.txt
-- Fixed ralph-loop.sh to handle table-format TODOs
-[2026-01-22 21:52] RALPH loop started (max 50 iterations)
-[2026-01-22 21:53] RALPH loop started (max 50 iterations)
-[2026-01-22 21:53] RALPH loop started (max 50 iterations)
-[2026-01-22 21:53] RALPH loop started (max 50 iterations)
-[2026-01-22 21:54] RALPH loop started (max 50 iterations)
-[2026-01-22 21:54] RALPH loop started (max 50 iterations)
-[2026-01-22 21:54] RALPH loop started (max 50 iterations)
-[2026-01-22 21:55] RALPH loop started (max 50 iterations)
-[2026-01-22 21:55] RALPH loop started (max 50 iterations)
-[2026-01-22 21:56] RALPH loop started (max 50 iterations)
-[2026-01-22 22:01] RALPH loop started (max 50 iterations)
-[2026-01-22 22:04] RALPH loop started (max 50 iterations)
-[2026-01-22 22:04] Iteration 1 had errors
-[2026-01-22 22:08] RALPH loop started (max 50 iterations)
-[2026-01-22 22:08] Iteration 1 had errors
-[2026-01-22 22:08] Iteration 2 had errors
-[2026-01-22 22:09] RALPH loop started (max 50 iterations)
-[2026-01-22 22:12] Iteration 1 had errors
-[2026-01-22 22:12] Iteration 2 had errors
-[2026-01-22 22:12] Iteration 3 had errors
-[2026-01-22 22:12] Iteration 4 had errors
-[2026-01-22 22:12] Iteration 5 had errors
-[2026-01-22 22:13] Iteration 6 had errors
-[2026-01-22 22:13] Iteration 7 had errors
-[2026-01-22 22:13] Iteration 8 had errors
-[2026-01-22 22:13] Iteration 9 had errors
-[2026-01-22 22:13] Iteration 10 had errors
-[2026-01-22 22:13] Iteration 11 had errors
-[2026-01-22 22:13] Iteration 12 had errors
-[2026-01-22 22:13] Iteration 13 had errors
-[2026-01-22 22:13] Iteration 14 had errors
-[2026-01-22 22:13] Iteration 15 had errors
-[2026-01-22 22:13] Iteration 16 had errors
-[2026-01-22 22:13] Iteration 17 had errors
-[2026-01-22 22:13] Iteration 18 had errors
-[2026-01-22 22:14] Iteration 19 had errors
-[2026-01-22 22:14] Iteration 20 had errors
-[2026-01-22 22:14] Iteration 21 had errors
-[2026-01-22 22:14] Iteration 22 had errors
-[2026-01-22 22:14] Iteration 23 had errors
-[2026-01-22 22:14] Iteration 24 had errors
-[2026-01-22 22:14] Iteration 25 had errors
-[2026-01-22 22:14] Iteration 26 had errors
-[2026-01-22 22:14] Iteration 27 had errors
-[2026-01-22 22:14] Iteration 28 had errors
-[2026-01-22 22:14] Iteration 29 had errors
-[2026-01-22 22:14] Iteration 30 had errors
-[2026-01-22 22:14] Iteration 31 had errors
-[2026-01-22 22:14] Iteration 32 had errors
-[2026-01-22 22:24] RALPH loop started (max 50 iterations)
-[2026-01-22T22:31:12Z] RALPH loop started (max 50 iterations)
-[2026-01-22T22:34:20Z] RALPH loop started (max 50 iterations)
