--- .ralph/status.json
@@ -1,10 +1,0 @@
-{
-  "task": "workspace-migration",
-  "status": "running",
-  "iteration": 1,
-  "maxIterations": 50,
-  "startedAt": "2026-01-22T21:24:00Z",
-  "lastRunAt": "2026-01-22T22:34:20Z",
-  "completedTasks": 44,
-  "totalTasks": 55
-}
