--- cli/Cargo.toml
@@ -1,60 +1,0 @@
-[package]
-name = "agent-tui"
-version = "0.3.6"
-edition = "2021"
-description = "CLI tool for AI agents to interact with TUI applications"
-license = "MIT"
-repository = "https://github.com/pproenca/agent-tui"
-homepage = "https://github.com/pproenca/agent-tui"
-documentation = "https://github.com/pproenca/agent-tui#readme"
-readme = "../README.md"
-keywords = ["tui", "terminal", "ai", "agent", "automation"]
-categories = ["command-line-utilities", "development-tools"]
-exclude = [
-    "tests/*",
-    ".github/*"
-]
-
-[lib]
-name = "agent_tui"
-path = "src/lib.rs"
-
-[[bin]]
-name = "agent-tui"
-path = "src/main.rs"
-
-[dependencies]
-clap = { version = "4.5", features = ["derive", "env"] }
-clap_complete = "4.5"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tokio = { version = "1.36", features = ["full"] }
-thiserror = "1.0"
-uuid = { version = "1.7", features = ["v4"] }
-rand = "0.8"
-chrono = { version = "0.4", features = ["serde"] }
-# For native terminal attach
-crossterm = "0.28"
-base64 = "0.22"
-# Pure Rust PTY + terminal emulation (no daemon)
-portable-pty = "0.8"
-vt100 = "0.15"
-regex = "1.10"
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2"
-
-[dev-dependencies]
-assert_cmd = "2.0"        # Binary testing - invoke CLI as subprocess
-predicates = "3.0"        # Assertion building for CLI output verification
-tempfile = "3.0"          # Isolated socket paths for mock daemon
-
-# cargo-udeps false positives (rand is used in connection.rs for jitter)
-[package.metadata.cargo-udeps.ignore]
-normal = ["rand"]
-
-[profile.release]
-opt-level = 3
-lto = true
-codegen-units = 1
-strip = true
