--- cli/Cross.toml
@@ -1,10 +1,0 @@
-[build.env]
-passthrough = [
-    "CARGO_TERM_COLOR",
-]
-
-[target.aarch64-unknown-linux-gnu]
-image = "ghcr.io/cross-rs/aarch64-unknown-linux-gnu:main"
-
-[target.x86_64-unknown-linux-gnu]
-image = "ghcr.io/cross-rs/x86_64-unknown-linux-gnu:main"
