--- cli/rustfmt.toml
@@ -1,7 +1,0 @@
-edition = "2024"
-format_code_in_doc_comments = true
-group_imports = "StdExternalCrate"
-imports_granularity = "Item"
-merge_derives = false
-style_edition = "2024"
-use_field_init_shorthand = true
