--- cli/src/attach.rs
@@ -1,229 +1,0 @@
-use crate::client::DaemonClient;
-use crate::color::Colors;
-use base64::{engine::general_purpose::STANDARD, Engine};
-use crossterm::{
-    event::{self, Event, KeyCode, KeyModifiers},
-    terminal::{self, disable_raw_mode, enable_raw_mode},
-};
-use serde_json::json;
-use std::io::{self, Write};
-use std::time::Duration;
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum AttachError {
-    #[error("Terminal error: {0}")]
-    Terminal(#[from] io::Error),
-}
-
-pub fn attach_ipc(client: &mut DaemonClient, session_id: &str) -> Result<(), AttachError> {
-    eprintln!(
-        "{} Attaching to session {}...",
-        Colors::dim("[attach]"),
-        Colors::session_id(session_id)
-    );
-    eprintln!(
-        "{} Press {} to detach.",
-        Colors::success("Connected!"),
-        Colors::bold("Ctrl+\\")
-    );
-    eprintln!();
-
-    enable_raw_mode().map_err(AttachError::Terminal)?;
-
-    let (cols, rows) = terminal::size().map_err(AttachError::Terminal)?;
-    let resize_params = json!({
-        "cols": cols,
-        "rows": rows,
-        "session": session_id
-    });
-    let _ = client.call("resize", Some(resize_params));
-
-    let result = attach_ipc_loop(client, session_id);
-
-    let _ = disable_raw_mode();
-
-    eprintln!();
-    eprintln!(
-        "{} Detached from session {}",
-        Colors::dim("[attach]"),
-        Colors::session_id(session_id)
-    );
-
-    result
-}
-
-fn attach_ipc_loop(client: &mut DaemonClient, session_id: &str) -> Result<(), AttachError> {
-    let mut stdout = io::stdout();
-
-    loop {
-        if event::poll(Duration::from_millis(10)).unwrap_or(false) {
-            match event::read() {
-                Ok(Event::Key(key_event)) => {
-                    if key_event.modifiers.contains(KeyModifiers::CONTROL)
-                        && key_event.code == KeyCode::Char('\\')
-                    {
-                        break;
-                    }
-
-                    if let Some(bytes) = key_event_to_bytes(&key_event) {
-                        let data_b64 = STANDARD.encode(&bytes);
-                        let params = json!({
-                            "session": session_id,
-                            "data": data_b64
-                        });
-                        if client.call("pty_write", Some(params)).is_err() {
-                            break;
-                        }
-                    }
-                }
-                Ok(Event::Resize(cols, rows)) => {
-                    let params = json!({
-                        "cols": cols,
-                        "rows": rows,
-                        "session": session_id
-                    });
-                    let _ = client.call("resize", Some(params));
-                }
-                Ok(_) => {}
-                Err(_) => break,
-            }
-        }
-
-        let read_params = json!({
-            "session": session_id,
-            "timeout_ms": 50
-        });
-        match client.call("pty_read", Some(read_params)) {
-            Ok(result) => {
-                if let Some(data_b64) = result.get("data").and_then(serde_json::Value::as_str) {
-                    if let Ok(data) = STANDARD.decode(data_b64) {
-                        if !data.is_empty() {
-                            if stdout.write_all(&data).is_err() {
-                                break;
-                            }
-                            if stdout.flush().is_err() {
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-            Err(_) => {
-                break;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn key_event_to_bytes(key_event: &event::KeyEvent) -> Option<Vec<u8>> {
-    use KeyCode::*;
-
-    let ctrl = key_event.modifiers.contains(KeyModifiers::CONTROL);
-    let alt = key_event.modifiers.contains(KeyModifiers::ALT);
-
-    match key_event.code {
-        Char(c) => {
-            if ctrl {
-                if c.is_ascii_lowercase() {
-                    Some(vec![c as u8 - b'a' + 1])
-                } else if c.is_ascii_uppercase() {
-                    Some(vec![c as u8 - b'A' + 1])
-                } else {
-                    match c {
-                        '[' | '3' => Some(vec![0x1b]),
-                        '\\' | '4' => Some(vec![0x1c]),
-                        ']' | '5' => Some(vec![0x1d]),
-                        '^' | '6' => Some(vec![0x1e]),
-                        '_' | '7' => Some(vec![0x1f]),
-                        '?' | '8' => Some(vec![0x7f]),
-                        ' ' | '2' | '@' => Some(vec![0x00]),
-                        _ => None,
-                    }
-                }
-            } else if alt {
-                Some(vec![0x1b, c as u8])
-            } else {
-                let mut buf = [0u8; 4];
-                let s = c.encode_utf8(&mut buf);
-                Some(s.as_bytes().to_vec())
-            }
-        }
-        Enter => Some(vec![b'\r']),
-        Tab => {
-            if key_event.modifiers.contains(KeyModifiers::SHIFT) {
-                Some(vec![0x1b, b'[', b'Z'])
-            } else {
-                Some(vec![b'\t'])
-            }
-        }
-        Backspace => Some(vec![0x7f]),
-        Delete => Some(vec![0x1b, b'[', b'3', b'~']),
-        Esc => Some(vec![0x1b]),
-        Up => Some(vec![0x1b, b'[', b'A']),
-        Down => Some(vec![0x1b, b'[', b'B']),
-        Right => Some(vec![0x1b, b'[', b'C']),
-        Left => Some(vec![0x1b, b'[', b'D']),
-        Home => Some(vec![0x1b, b'[', b'H']),
-        End => Some(vec![0x1b, b'[', b'F']),
-        PageUp => Some(vec![0x1b, b'[', b'5', b'~']),
-        PageDown => Some(vec![0x1b, b'[', b'6', b'~']),
-        Insert => Some(vec![0x1b, b'[', b'2', b'~']),
-        F(n) => {
-            let seq = match n {
-                1 => vec![0x1b, b'O', b'P'],
-                2 => vec![0x1b, b'O', b'Q'],
-                3 => vec![0x1b, b'O', b'R'],
-                4 => vec![0x1b, b'O', b'S'],
-                5 => vec![0x1b, b'[', b'1', b'5', b'~'],
-                6 => vec![0x1b, b'[', b'1', b'7', b'~'],
-                7 => vec![0x1b, b'[', b'1', b'8', b'~'],
-                8 => vec![0x1b, b'[', b'1', b'9', b'~'],
-                9 => vec![0x1b, b'[', b'2', b'0', b'~'],
-                10 => vec![0x1b, b'[', b'2', b'1', b'~'],
-                11 => vec![0x1b, b'[', b'2', b'3', b'~'],
-                12 => vec![0x1b, b'[', b'2', b'4', b'~'],
-                _ => return None,
-            };
-            Some(seq)
-        }
-        _ => None,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_key_event_to_bytes_char() {
-        let event = event::KeyEvent::new(KeyCode::Char('a'), KeyModifiers::NONE);
-        assert_eq!(key_event_to_bytes(&event), Some(vec![b'a']));
-    }
-
-    #[test]
-    fn test_key_event_to_bytes_ctrl() {
-        let event = event::KeyEvent::new(KeyCode::Char('c'), KeyModifiers::CONTROL);
-        assert_eq!(key_event_to_bytes(&event), Some(vec![0x03]));
-    }
-
-    #[test]
-    fn test_key_event_to_bytes_enter() {
-        let event = event::KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE);
-        assert_eq!(key_event_to_bytes(&event), Some(vec![b'\r']));
-    }
-
-    #[test]
-    fn test_key_event_to_bytes_arrow() {
-        let event = event::KeyEvent::new(KeyCode::Up, KeyModifiers::NONE);
-        assert_eq!(key_event_to_bytes(&event), Some(vec![0x1b, b'[', b'A']));
-    }
-
-    #[test]
-    fn test_key_event_to_bytes_f1() {
-        let event = event::KeyEvent::new(KeyCode::F(1), KeyModifiers::NONE);
-        assert_eq!(key_event_to_bytes(&event), Some(vec![0x1b, b'O', b'P']));
-    }
-}
