--- cli/src/daemon/ansi_keys.rs
@@ -1,6 +1,0 @@
-//! ANSI escape sequences for cursor/arrow keys.
-
-pub const UP: &[u8] = b"\x1b[A";
-pub const DOWN: &[u8] = b"\x1b[B";
-pub const RIGHT: &[u8] = b"\x1b[C";
-pub const LEFT: &[u8] = b"\x1b[D";
