--- cli/src/daemon/mod.rs
@@ -1,8 +1,0 @@
-mod ansi_keys;
-mod error_messages;
-mod lock_helpers;
-mod rpc_types;
-mod select_helpers;
-pub mod server;
-
-pub use server::{socket_path, start_daemon};
