--- cli/src/lib.rs
@@ -1,14 +1,0 @@
-pub mod pty;
-pub mod session;
-pub mod sync_utils;
-pub mod terminal;
-pub mod vom;
-pub mod wait;
-
-pub use pty::{PtyError, PtyHandle};
-pub use session::{
-    Element, ElementType, Position, Session, SessionError, SessionId, SessionInfo, SessionManager,
-};
-pub use terminal::{CursorPosition, ScreenBuffer, VirtualTerminal};
-pub use vom::{Component, Role};
-pub use wait::WaitCondition;
