+++ crates/agent-tui/Cargo.toml
@@ -34,3 +34,5 @@ base64.workspace = true
 assert_cmd.workspace = true
 predicates.workspace = true
 tempfile.workspace = true
+tokio.workspace = true
+uuid.workspace = true
