+++ crates/agent-tui/tests/common/mock_daemon.rs
