+++ crates/agent-tui/tests/common/mod.rs
