+++ crates/agent-tui/tests/e2e_daemon_tests.rs
@@ -14,7 +14,7 @@
 
 mod common;
 
-use common::{TestHarness, TEST_COLS, TEST_ROWS, TEST_SESSION_ID};
+use common::{TEST_COLS, TEST_ROWS, TEST_SESSION_ID, TestHarness};
 use predicates::prelude::*;
 use serde_json::json;
